requirements.txt
pandas
numpy
scipy
scikit-learn
boto3
python-dotenv
python-dateutil
pytz
matplotlib
resend
